# docs/05-api-documentation/openapi-spec.yaml
openapi: 3.1.0
info:
  title: Social Media AI - Content Distribution API
  version: 1.0.0
  description: |
    Complete API for AI-powered social media content creation, scheduling,
    analytics, and multi-platform distribution.
  contact:
    name: API Support
    email: api-support@contentdistribution.ai
  license:
    name: Proprietary
    url: https://contentdistribution.ai/license
servers:
  - url: https://api.contentdistribution.ai/v1
    description: Production
  - url: https://staging-api.contentdistribution.ai/v1
    description: Staging
security:
  - BearerAuth: []
...
paths:
...
  /posts:
    post:
      tags: [Posts]
      summary: Create a post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'

    get:
      tags: [Posts]
      summary: List posts
      parameters:
        - name: workspaceId
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: sort
          in: query
          schema:
            type: string
            default: -createdAt
        - name: filter[status]
          in: query
          schema:
            type: string
            enum: [draft, scheduled, publishing, published, failed]
        - name: filter[platform]
          in: query
          schema:
            type: string
        - name: filter[category]
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Posts list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedPosts'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Pagination cursor
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
      description: Results per page
...
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        content:
          type: string
        platforms:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, scheduled, publishing, published, failed]
        media:
          type: array
          items:
            type: object
        tags:
          type: array
          items:
            type: string
        category:
          type: string
        scheduledAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
        createdBy:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePostRequest:
      type: object
      required: [workspaceId, content, platforms]
      properties:
        workspaceId:
          type: string
        content:
          type: string
          minLength: 1
          maxLength: 10000
        platforms:
          type: array
          minItems: 1
          items:
            type: string
            enum: [twitter, linkedin, facebook, instagram, tiktok, youtube, pinterest]
        mediaIds:
          type: array
          items:
            type: string
        scheduledAt:
          type: string
          format: date-time
        tags:
          type: array
          maxItems: 20
          items:
            type: string
        category:
          type: string
        variables:
          type: object
...
    PaginatedPosts:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        cursor:
          type: string
          nullable: true
        hasMore:
          type: boolean
        total:
          type: integer